set up concurrency
