+++ hw-3/bombard/src/commands/run/index.ts
@@ -1,11 +1,5 @@
 import { Command, Flags } from "@oclif/core";
 
-interface IStats {
-  success: number;
-  fail: number;
-  avg: number;
-}
-
 export default class Run extends Command {
   static description = "Say hello";
 
@@ -43,28 +37,74 @@ hello friend from oclif! (./src/commands/hello/index.ts)
   ];
 
   async run(): Promise<void> {
-    const timeZero = Date.now();
-    const { args, flags } = await this.parse(Hello);
+    const { args, flags } = await this.parse(Run);
     const http = require("http");
+    const async = require("async");
 
     const reqCount = flags.requests ?? 10;
     const concurr = flags.concurrency ?? 4;
-    const withBody = flags.body;
+    // const withBody = flags.body;
 
     const url = args.url;
+    const requests = [];
+    const zeroTime = Date.now();
+
+    for (let i = 0; i < reqCount; i += 1) {
+      requests.push((callback: any) => {
+        const startTime = Date.now();
+
+        http
+          .get(url, (res: any) => {
+            let data = "";
+
+            // A chunk of data has been received.
+            res.on("data", (chunk: any) => {
+              data += chunk;
+            });
+
+            // The whole response has been received. Print out the result.
+            res.on("end", () => {
+              const endTime = Date.now();
+
+              callback(null, {
+                call: startTime - zeroTime,
+                duration: endTime - startTime,
+                data: data,
+              });
+            });
+          })
+          .on("error", (error: any) => {
+            callback(null, { error: error });
+          });
+      });
+    }
 
     let success = 0;
     let fail = 0;
     let avg = 0;
 
-    await bombard(reqCount, concurr, url);
+    async.parallelLimit(requests, concurr, (err: any, results: any) => {
+      if (err) {
+        this.log("Some error occurred");
+      } else {
+        for (const result of results) {
+          const { duration, error } = result;
 
-    this.log(`bombarded ${reqCount} times`);
-    this.log(`${success} successful, ${fail} failing`);
+          if (error) {
+            fail += 1;
+          } else {
+            success += 1;
+            avg += duration;
+          }
+        }
 
-    if (success > 0) {
-      avg /= success;
-      this.log(`average response time ${avg / success}ms`);
-    }
+        this.log(`bombarded ${reqCount} times`);
+        this.log(`${success} successful, ${fail} failing`);
+
+        if (success > 0) {
+          this.log(`average response time ${avg / success}ms`);
+        }
+      }
+    });
   }
 }
